--- src/hybrid/compositor/index.js
@@ -1,1 +1,0 @@
-import './compositor'
