+++ src/index.js
@@ -32,6 +32,6 @@ import './systems'; // custom AFRAME systems
 import './geometries'; // custom AFRAME geometries
 import './components'; // custom AFRAME components
 import './postprocessing'; // post-processing
-import './hybrid'; // hybrid rendering
+import './render-fusion'; // hybrid rendering
 import './webxr'; // special handler for webxr devices
 import './webar'; // special handler for non-webxr devices
