+++ src/postprocessing/index.js
@@ -1,13 +1,1 @@
-import {EffectComposer} from './effect-composer';
-import {GlitchPass} from './passes/glitch-pass';
-import {BloomPass} from './passes/bloom-pass';
-import {UnrealBloomPass} from './passes/unreal-bloom-pass';
-import {ShaderPass} from './passes/shader-pass';
-
-export {
-    EffectComposer,
-    GlitchPass,
-    BloomPass,
-    UnrealBloomPass,
-    ShaderPass
-}
+import './effects';
