+++ src/render-fusion/index.js
@@ -1,3 +1,3 @@
-import './render-client';
 import './compositor';
+import './render-client';
 import './remote-render';
