+++ src/render-fusion/webrtc-stats.js
